Rework Layers and move into structure file
